requirements.txt
Django==3.1.1
Djangorestframework==3.1.1
mysqlclient-1.4.6-cp38-cp38-win32.whl